hack: fix dlv installation
the way that we're currently asking `dlv` to be installed (via `go get`
in a directory that is module aware), it's as if we were making dlv a
dependency of concourse, which is not really what we want.

as explained in an issue in dlv's repo (see [1]), this can be fixed by
turning go modules off during the installation.

[1]: https://github.com/go-delve/delve/issues/1991#issuecomment-609706835

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>