atc: behaviour: fix flag parsing
concourse/concourse#5421

This was tricky to track down and involved some use of dlv, but the test
probably explains itself - we were vaidating but not actually unmarshalling
before.

The reason we weren't catching this is because we were missing an integration
test across the entire parsing -> running -> HTTP request boundary. We had
an integration test on an unhappy path, but the ATC integration test didn't
actually exercise the flag parsing.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>