topgun/k8s: fix duration format
Now that we're using stricter env var parsing, these syntax errors are causing
k8s-topgun failures. In fact, this means that the properties were never getting
parsed properly, and therefore that the cache duration was effectively the
default at the binary level, which, at the time of writing, is 1 minute.

Let's keep an eye out for whether this change increases the running time of
k8s-topgun.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>