web: behaviour: recover from turbulence view
If any network request fails once, the dashboard will be stuck in the
"show turbulence" page view forever (until you refresh the page). This
happens fairly frequently to me at least - if you have the dashboard
open, close your laptop for a little, and come back, the page will be
unusable due to the turbulence view.

This commit allows the turbulence view to disappear if all the once
failing requests start succeeding. It no longer considers the "team
pipelines" endpoint response for displaying the show turbulence view -
if it ever errors, we'd still be stuck in the turbulence view forever
(since nothing would trigger another `FetchPipelines` effect). It's also
not that valuable - if there is truly turbulence, the next network tick
(in < 5 seconds) will display this.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>