topgun: fix creds mgmt tests
there are a set of tests in that suite that requires the user to provide
their own namespaces for the teams that the `web` nodes are supposed to
retrieve credentials from:

	1. create namespace
	2. tell concourse that team & pipeline credentials should come
	from that specific namespace
	3. put secrets in that namespace
	4. have `web` nodes reading secrets from that namespace

now, despite the user having the ability to do all of those first 3
steps, the 4th is something that the `web` node on its own, using its
own credentials (by means of it serviceAccount token).

internally, kubernetes will check if that serviceAccount can do what it
wants (fetch secrets from that namespace) by verifying which roles are
bound to it.

previously, the thing (rolebinding) that would bind the role (ability to
read secrets from the namespace) to the actor (serviceAccount) was
created by default by the chart, but we changed that (see note 1) to be
more flexible.

this means that the test should reflect the new need of the user
providing the rolebinding.

note 1: https://github.com/concourse/concourse-chart/pull/4

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>