web: behaviour: stop polling ListAllJobs on 501
concourse/concourse#5470

Note that the only way to get it to start polling again would be to reconfigure
the server _and_ refresh the page. I think that's pretty reasonable (bouncing
the server is probably going to induce downtime/http errors anyway).

Should we ever need to re-start polling after a 501 -- we'll cross that bridge
when we come to it.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>