wats: run watsjs in serial
From looking at the `failure.png` of one of the builds that failed on CI
for the "shows pipelines in their correct order" test, I noticed that
the team that was on the screen was not the same team that the test was
running in. I suspect the following happened:

1. "shows pipelines in their correct order" starts
2. It scrolls the team for that test into view so that pipeline cards
   will be rendered
3. Another test starts and creates a new team. The scroll position of
   the document remains, but the team we care about is now offscreen
4. Lazy rendering removes the relevant pipeline cards from the document

One option to fix this is to try to constantly scroll the team into
view, rather than just once at the start. Another option would be to try
to increase the viewport size so that all pipelines are rendered (I
tried this with the `--window-size` puppeteer flag to no avail). This
commit takes a different approach - just run the tests in serial to
ensure the document doesn't change from under us while we're running
this test. I can imagine there being other unexpected test bugs from
running all the tests in parallel that are hard to reason about, and the
tests aren't that slow (certainly, watsjs will still be much faster than
the unit suite).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>