atc,tf: structure: robustify active users test
concourse/concourse#5628

I originally wrote this test as a throwaway during a discussion about an issue I
wasn't actively working onand didn't expect it to stick around in exactly the
same form -- famous last words.

I was able to consistently reproduce the testflight flake mentioned in the
linked issue by always starting with an empty database. Initially I rewrote the
test as an atc integration test just to make it easier to attach a debugger --
I was finding the
`docker-compose down -v && docker-compose up -d && hack/trace web` dance a bit
clunky. Conveniently once I did, the flake became completely reproducible
because every atc integration test starts with a fresh database.

I didn't really thoroughly investigate, but I strongly suspect the flake came
from the fact that there is an asynchronous `batcher` that is responsible for
updating active user data - so there can be a delay between logging in and when
the users table gets updated.

On the one hand, it may be worth asking "why doesn't the database get
synchronously updated?" and the answer is, roughly, "optimization reasons".
Allowing for that, then, presumably there are some thorough unit tests for this
carefully-optimized procedure. So why have an integration test at all if there's
some non-determinism?

Because this feature is really, really important. It gets used to define billing
in some cases. So it's worth having a potentially slow-running and complex
integration test if it will verify that billing still works properly.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>