containerd: modify RootFSManager interface
Add LookupUser to RootFSManager's interface. Replace specs.Spec with
the path to rootfs in the signature for SetupCwd and LookupUser as this
is all it requires from the Spec object. Passing the entire object makes
it less clear while reading from users of RootFSManager what it does with Spec.

Regenerate fakes for RootFSManager.

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>