atc: check if version exists during alg compute
When we determine a set of versions within the algorithm, we should only
be using versions that exists in the resource config versions table. For
the pinned and individual resolvers, we compute our version for the
input using the resource config versions table so we will only compute a
version that already exists. For the group resolver, we use the
successful build outputs table so we might end up computing a version
that exists in the successful build outputs but not in the resource
config versions table.

In order to avoid the case of us picking a version that does not exist
in the group resolver, we added a check when we set a new version as a
version candidate. The check will query the resource config versions
table for that version to ensure that the version exists before we set
the candidate to that version.

[#5655]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: James Thomson <jthomson@pivotal.io>