atc: move waiting on limiter before spawning goroutine
This way we won't have many goroutines all waiting on the rate limiter,
and rather have the goroutine only spawn after it is able to get past
the rate limit. Also, it will make the timing during the span context
for checker.Run be more accurate because it no longer includes the time
it takes to wait for the limiter.

[#4624]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>