atc: do not rate limit manual triggered checks
When we added the rate limiting to the number of checks that can be
started, the default rate limit is calculated by figuring out the rate
at which checks should happen in order to check all the resources within
the resource checking interval. This calculation does not take into
account checks that are manually triggered. This means that if the rate
is set low enough and a user manually triggers a check, that check might
be slow to run. That is probably unwanted behaviour, since if a user
runs a check it should probably be run as soon as possible.

This commit makes it so that manually triggered checks will not be rate
limited.

[#4624]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>