web: structure: standardize font weights/families
This commit attempts to bring some order to the chaos that is style
constants (at least around typography). I was afraid to mess with the
`.less` files too much - some of it seems legacy/unused, but some of it
seems important.

The motivation for this is to add a new standard font weight -
`font-weight: 900`.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>