atc: structure: replace placeholder name
I wanted to call this PlanConfig but I couldn't until the old PlanConfig
was removed. It was previously called Step.

So... here's the whole sequence of what happened:

* PlanConfig -> (removed)
* Step -> PlanConfigXXX
* new Step type introduced to replace old PlanConfig type

This is admittedly a really awkward naming dance, but I think it makes a
little more sense this way because now NewPlan takes a PlanConfig to
return a Plan.

Signed-off-by: Alex Suraci <asuraci@pivotal.io>