Revert "docker-compose: break worker/web dependency"
this breaks downgrade tests; the worker can't unregister in time, since
the web node just goes away immediately, and workers give up when all
web nodes are unreachable, assuming the cluster is being destroyed.

as a result, the check container ends up sticking around on a stalled
worker, and the check eventually times out. a better solution might be
to have checking not attempt to use a container that's on a stalled
worker, but let's not block the pipeline on that in the meantime.

This reverts commit f2237d3a4b3c04855bd5fa6af1d372b8d22bc57a.